# Conflicts:
#	.idea/dbnavigator.xml
#	.idea/deploymentTargetSelector.xml
#	.idea/misc.xml

+++ app/src/main/java/com/zinhao/kikoeru/MusicChangeListener.java
@@ -3,7 +3,7 @@ package com.zinhao.kikoeru;
 import org.json.JSONObject;
 
 public interface MusicChangeListener {
-    void onAlbumChange(int rjNumber);
+    void onAlbumChange(long rjNumber);
 
     void onAudioChange(JSONObject audio);
